+++ webpack/serve.js
@@ -9,7 +9,6 @@ const applyWebpackOverrides = require('./apply-webpack-overrides');
 const logger = require('./logger')();
 const buildConfig = require('../webpack.config.cli.dev');
 const sharedStylesConfig = require('../webpack.config.cli.shared.styles');
-const buildServiceWorkerConfig = require('../webpack.config.service.worker');
 
 const cwd = path.resolve();
 const platformModulePath = path.join(cwd, 'node_modules');
@@ -24,13 +23,6 @@ module.exports = function serve(stripesConfig, options) {
   return new Promise((resolve) => {
     logger.log('starting serve...');
     const app = express();
-
-    // service worker config
-    // update resolve/resolveLoader in order to find the micro-stripes-config
-    // virtual module configured by buildServiceWorkerConfig()
-    const serviceWorkerConfig = buildServiceWorkerConfig(stripesConfig);
-    serviceWorkerConfig.resolve = { modules: ['node_modules', platformModulePath, coreModulePath] };
-    serviceWorkerConfig.resolveLoader = { modules: ['node_modules', platformModulePath, coreModulePath] };
 
     let config = buildConfig(stripesConfig);
 
@@ -49,10 +41,9 @@ module.exports = function serve(stripesConfig, options) {
     config = applyWebpackOverrides(options.webpackOverrides, config);
 
     logger.log('assign final webpack config', config);
-    const compiler = webpack([serviceWorkerConfig, config]);
-    const [swCompiler, stripesCompiler] = compiler.compilers;
+    const compiler = webpack(config);
 
-    stripesCompiler.hooks.done.tap('StripesCoreServe', stats => resolve(stats));
+    compiler.hooks.done.tap('StripesCoreServe', stats => resolve(stats));
 
     const port = options.port || process.env.STRIPES_PORT || 3000;
     const host = options.host || process.env.STRIPES_HOST || 'localhost';
@@ -65,10 +56,6 @@ module.exports = function serve(stripesConfig, options) {
     // https://github.com/bripkens/connect-history-api-fallback/blob/master/examples/static-files-and-index-rewrite
     app.use(staticFileMiddleware);
 
-    app.use(webpackDevMiddleware(swCompiler, {
-      publicPath: serviceWorkerConfig.output.publicPath,
-    }));
-
     // Process index rewrite before webpack-dev-middleware
     // to respond with webpack's dist copy of index.html
     app.use(connectHistoryApiFallback({
@@ -76,11 +63,11 @@ module.exports = function serve(stripesConfig, options) {
       htmlAcceptHeaders: ['text/html', 'application/xhtml+xml'],
     }));
 
-    app.use(webpackDevMiddleware(stripesCompiler, {
+    app.use(webpackDevMiddleware(compiler, {
       publicPath: config.output.publicPath,
     }));
 
-    app.use(webpackHotMiddleware(stripesCompiler));
+    app.use(webpackHotMiddleware(compiler));
 
     app.listen(port, host, (err) => {
       if (err) {